refactor: split workspace_temp handler into focused modules
Move workspace creation into operations/create.rs, editor/hook logic
into operations/editor.rs, and status-update hook into hooks.rs,
leaving handler.rs as thin orchestration only. Closes #398.
