+++ .centy/issues/d2d796fb-ad17-4e43-8a15-3035e3f8c2f7.md
@@ -1,10 +1,10 @@
 ---
 # This file is managed by Centy. Use the Centy CLI to modify it.
 displayNumber: 398
-status: in-progress
+status: closed
 priority: 2
 createdAt: 2026-04-04T11:55:40.212144+00:00
-updatedAt: 2026-04-04T12:04:27.824541+00:00
+updatedAt: 2026-04-04T12:14:30.026246+00:00
 ---
 
 # Split workspace_temp handler — separate creation, editor, and hooks phases
