+++ src/server/handlers/workspace_temp/handler.rs
@@ -1,57 +1,15 @@
-use super::editor::{open_editor_with_hooks, try_update_status_on_open};
+use super::hooks::try_update_status_on_open;
+use super::operations::run_create_workspace;
 use super::response::err_response;
 use crate::config::read_config;
-use crate::item::entities::issue::Issue;
 use crate::registry::track_project_async;
 use crate::server::assert_service::assert_initialized;
 use crate::server::proto::{OpenInTempWorkspaceResponse, OpenInTempWorkspaceWithEditorRequest};
 use crate::server::resolve::resolve_issue;
 use crate::server::structured_error::{to_error_json, StructuredError};
-use crate::workspace::{create_temp_workspace, CreateWorkspaceOptions};
 use std::path::Path;
 use tonic::{Response, Status};
 
-async fn run_create_workspace(
-    project_path: &Path,
-    req_project_path: &str,
-    req_editor_id: &str,
-    issue: Issue,
-) -> Result<Response<OpenInTempWorkspaceResponse>, Status> {
-    match create_temp_workspace(CreateWorkspaceOptions {
-        source_project_path: project_path.to_path_buf(),
-        issue: issue.clone(),
-    })
-    .await
-    {
-        Ok(result) => {
-            let workspace_path = result.workspace_path.to_string_lossy().to_string();
-            let editor_opened = open_editor_with_hooks(
-                req_editor_id,
-                &workspace_path,
-                issue.metadata.display_number,
-                project_path,
-            );
-            Ok(Response::new(OpenInTempWorkspaceResponse {
-                success: true,
-                error: String::new(),
-                workspace_path,
-                issue_id: issue.id.clone(),
-                display_number: issue.metadata.display_number,
-                expires_at: String::new(),
-                editor_opened,
-                requires_status_config: false,
-                workspace_reused: result.workspace_reused,
-                original_created_at: String::new(),
-            }))
-        }
-        Err(e) => Ok(err_response(
-            to_error_json(req_project_path, &e),
-            String::new(),
-            0,
-            false,
-        )),
-    }
-}
 pub async fn open_in_temp_workspace(
     req: OpenInTempWorkspaceWithEditorRequest,
 ) -> Result<Response<OpenInTempWorkspaceResponse>, Status> {
@@ -83,10 +41,14 @@ pub async fn open_in_temp_workspace(
     if requires_status_config {
         return Ok(err_response(
             StructuredError::new(
-                &req.project_path, "STATUS_CONFIG_REQUIRED",
+                &req.project_path,
+                "STATUS_CONFIG_REQUIRED",
                 "Status update preference not configured. Set workspace.updateStatusOnOpen in your project config.".to_string(),
-            ).to_json(),
-            issue.id.clone(), issue.metadata.display_number, true,
+            )
+            .to_json(),
+            issue.id.clone(),
+            issue.metadata.display_number,
+            true,
         ));
     }
     try_update_status_on_open(
