+++ src/server/handlers/workspace_temp/mod.rs
@@ -1,4 +1,5 @@
-mod editor;
 mod handler;
+mod hooks;
+mod operations;
 mod response;
 pub use handler::open_in_temp_workspace;
